refactor: 경로 재설정 및, 테스트 코드에 맞춘 BridgeMaker 위치 재정의
